splits views into multiple files
