+++ server/pos/views/stock.py
@@ -1,22 +1,11 @@
 from django.shortcuts import get_object_or_404
-from django.utils import timezone
 
-from pos.models.shift import Shift
 from pos.models.stock import Category, Ingredient, Item, Order, OrderLine, Purchase
-from pos.models.user import User
 
-from pos.serializers.shift import ShiftSerializer
 from pos.serializers.stock import CategorySerializer, IngredientSerializer, ItemSerializer, OrderLineSerializer, OrderSerializer, PurchaseSerializer
-from pos.serializers.user import CustomerSerializer
 
 from rest_framework import viewsets
 from rest_framework.response import Response
-
-
-# ViewSets define the view behavior.
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = CustomerSerializer
 
 
 class IngredientViewSet(viewsets.ModelViewSet):
@@ -42,19 +31,6 @@ class ItemViewSet(viewsets.ModelViewSet):
 class OrderViewSet(viewsets.ModelViewSet):
     queryset = Order.objects.all()
     serializer_class = OrderSerializer
-
-
-class ShiftViewSet(viewsets.ModelViewSet):
-    queryset = Shift.objects.all()
-    serializer_class = ShiftSerializer
-
-    def create(self, request, *args, **kwargs):
-        if Shift.objects.count() > 0:
-            current_shift = Shift.objects.latest('id')
-            current_shift.end = timezone.now()
-            current_shift.save()
-            # TODO: Calculate how much money should be in the register
-        return super(ShiftViewSet, self).create(request, args, kwargs)
 
 
 class PurchaseViewSet(viewsets.ViewSet):
